remove publish from content list view
