+++ joplin/base/models/information_page.py
@@ -53,7 +53,7 @@ class InformationPage(JanisBasePage):
             "or",
             ConditionalPublishRequirement(
                 RelationPublishRequirement("related_departments"),
-                "and",
+                "or",
                 FieldPublishRequirement("coa_global"),
             ),
             message="You must have at least 1 topic or 1 department or 'Top Level' checked."
